Fix for a problem noticed by running MTT via ni...
Fix for a problem noticed by running MTT via nightly cron -- if you
don't --force or don't rm -rf the scratch dir between runs, even if
MTT got a new MPI source, it thought that it had installed it already,
and therefore didn't re-build/re-run any of the tests.  

The issue turned out to be that MTT only identified MPI's by their INI
file section name.  Hence, it had no concept that a different version
of the same section name is really a different MPI installation.  This
patch adds in version number tracking into the MTT meta data so that
if different versions of the same MPI INI section are avaiable, they
are recognized correctly by MTT as different MPI instances (and
therefore it needs to be tested, etc.).

Note that this is still technically a problem for the test get's -- if
we update tests but don't update the MPI version, the tests will not
run.  This is probably not a huge problem because we will be updating
the MPI every night.  But it probably should be fixed someday.

This commit was SVN r209.