First cut at a scriptlet after each mpirun whic...
First cut at a scriptlet after each mpirun which will guarantee that
all orteds are dead and all the session directories are removed.

Needs testing on various OS's to ensure that we get the right
killall/ps functionality depending on which OS you're running on.

This commit was SVN r213.