Refs #37
The mtt client now explicitly reports whether a test has timed out (to
differentiate it from failing to meet its "pass" criteria).  This is
helpful for sysadmins to adjust the timeout value in their ini files.

The timed_out value is reported in the results.  See ticket #37 for
details of the possible (pass, timed_out) tuple possibilities.

This commit was SVN r243.