Fixes #44.
This commit fixes up the client side.  submit/index.php is still not
in MTT SVN, so I committed the change over there in the ompi-www
repository.  All it does is look for a PING variable in the form, and
if it finds it, exits (successfully) before doing anything.

This commit was SVN r254.