Refs #42
Commit a compromise for #42 that will get people able to submit test
run data.  This is not the real fix.  The idea with this commit is
just to skip the two fields that we know are problematic.

In discussions with Ethan about #42 and #35, we have come to the
conclusion that #35 is going to end up revamping submit/index.php, so
there was no point in doing a real fix for #42 -- just get people able
to submit and let #35 put in the real fix for ignoring fields that it
doesn't understand.

So as of this commit, people can submit test run data again.  However,
I'm leaving #42 open so that the handy SQL for selcting the column
names is still available so that I can incorporate it into #35.

This commit was SVN r266.