Change how we check for non-existent INI file s...
Change how we check for non-existent INI file sections.  Load in 2
copies of the INI file and only filter one of them.  When checking for
non-existent sections in the phases (I chose not to duplicate all the
traversing / splitting logic in a separate subroutine just to warn for
non-existent sections), check in the un-filtered INI data.

Fixes #165, #166.

This commit was SVN r532.