Fixes #199
`highlight_match` run amuck again. Only wrap matches in
highlighted font tags if the match contains a word character
(`[a-zA-Z0-9_]`).  Consider how many pointless `<font>` tags
get printed on a NULL search pattern with 38,000 characters
of STDOUT/STDERR (for ~500 lines of output)...  With this
fix, a formerly 5.6MB report is now a 175KB report!  Yoinks.

This commit was SVN r616.