Add `mtt-submit` script for submitting results ...
Add `mtt-submit` script for submitting results to the MTT
database (This script used to just be posted on the MTT Trac
wiki).  Use case: for submitting a batch of test results
that were not submitted because the submitting node was
unable to access the Internet.  Options for mtt-submit are
as follows:

{{{
./mtt-submit
  -f    File to post
  -g    Glob the -f argument
        (quote the -f argument, e.g., -f '/foo/files*')
  -l    fileList (file containing a newline delimited list
        of files to POST)
  -s    Script to use
  -h    Host name to post to
  -u    HTTP Username
  -p    HTTP Password
  -d    Debug mode
}}}

CAUTION: Use this script with care!  For example, you could
end up with duplicate test results in the database if you
used mtt-submit twice on the same file.

This commit was SVN r635.