Fixes #230
It seems that on some platforms (e.g., RHEL4U3/x86-64), wrap() does
not deal well with Very Large strings.  This workaround simply limits
how large the strings are that we give to wrap().  If the string is
too large, we simply use print without wrap().

This commit was SVN r682.