Fixes #307.
There was a corner case where if displaying multiple results with
differing x axis values, the x xaxis was not displayed properly.  This
patch reduces the number of loops and copies surrounding xaxis
processing and makes a true "union" xaxis that represents a union of
all xaxis values used for all lines on a single graph.

This commit was SVN r1014.