Refs #295.
Add &test_alloc() funclet which returns the default value of "slot" or
the value of the alloc field in the test run section.  The intended
values are "slot" or "node", indicating that the MCW rank processes
should be allocated either by node or by slot.

Still testing the coresponding perf bakeoff template INI file for
using this funclet properly.

This commit was SVN r1015.