Fixes #303: fix the logic for determining if th...
Fixes #303: fix the logic for determining if there is performance data
available for a particular result or not.

Also put in a workaround for a warning that has been appearing a lot
lately: 

{{{
[client 65.14.168.110] PHP Notice:  Use of undefined constant undef - assumed 'undef' in /nfs/rontok/xraid/data/osl/www/www.open-mpi.org/mtt/submit/index.php(75): eval()'d code on line 49
}}}

Made a dummy variable named "undef" and set it equal to the empty
string.

Also removed a little extra logic for a case that cannot occur.

This commit was SVN r1028.