Stats are collected midnight to midnight.
Also if we have to guess an mpi_install_id or test_build_id because the client
did not submit a valid one then point it to an 'invalid' tuple. This keeps us
from corrupting the database by guessing wrong. Adjust the schema so that it
inserts 'undef' values at tuple 0 that we can point to.

This commit was SVN r1033.