Add a periodic pause so that we don't block the...
Add a periodic pause so that we don't block the database for too long at a time.
If we have run for > 60 seconds then disconnect and wait 10 seonds before 
continuing. This has proven effective in allowing testing to continue while
the database is being maintained.

This commit was SVN r1151.