Prevent MTT from executing the same SQL express...
Prevent MTT from executing the same SQL expression twice by allowing it
to reuse the resource returned by the previous execution.

This will only save a couple seconds since PostgreSQL will have cached the
first result, and if we are luckly will return it fairly quickly. If we are
unlucky then we payed the query price twice. :(

This commit was SVN r1154.