Change the "exclusive" field in test run sectio...
Change the "exclusive" field in test run sections: now the numeric
value represents a priority.  If the same test is listed in multiple
test run groups, it will end up in the group with the highest
"exclusive" value.  If the same test is listed in multiple groups with
the same exclusive value, it's non-deterministic as to which group it
remains in (which is the prior behavior anyway).

This commit was SVN r1292.