Allow new fields in the reporter section: skip_...
Allow new fields in the reporter section: skip_mpi_get and
skip_mpi_install.  This allows a reporter to elect to be skipped for
specific mpi_get and mpi_install sections -- handy for having a single
INI file that can be used for both internal and external testing.

This commit was SVN r1294.