Refs #70
Add experimental "new" failure reporter script: new-result-reporter.pl.
The script creates a static HTML page containing links to MTT results
that have changed (see below link for a sample).

 http://www.open-mpi.org/~emallove/new-result-reporter-preview.html

It bombards the database with queries, so - for testing purposes - the
above output HTML page was creating after running mtt-result-reporter.pl
for only 15 seconds, but on only a six hour period of MTT results
- it can crank away for a looong time.  (Fixing #362 would greatly
shrink the new-result-reporter.pl's output HTML pages.)

The script looks through a batch of test results (always delimited by
date range, and optionally delimited by other fields such as MPI
version).  Out of these results, it looks for pairs of test cases that
are similar, but had differing outcomes.  There is a range of
"similarness" that is illustrated in shades of red table cells.  The
"redder" the cell, the more similar the pair of test cases.
Specifically, "more similar" means more database columns match between
the pairs of tests.  It is assumed that if two test cases are nearly
identical, but have different outcomes, this would be more relevant to
the MPI source - and hence - these results are redder and at the top of
the page.  Right now, the script is generalized such that no columns are
weighted as more important than others, but we will need to hone in on
which matching columns (and groupings of columns) are most critical.
All the links displayed are unique.  new-result-reporter.pl needs
improvement in the following areas:

 1. Report fewer, and more meaningful links to results
 2. Optimize the algorithm used for finding result pairs, which might
    lessen the number of queries it needs to run (it is basically brute
    force right now)
 3. Optimize the SELECTs (using SQL magic like indexes, caching,
    temporary tables, etc.)

This commit was SVN r1359.