Per discussion on IM and on the mtt-devel list....
Per discussion on IM and on the mtt-devel list...

The old merge_hashes routine didn't cover all the corner cases.  So we
changed the orientation: instead of saving the entire hash in each
unique file and then figuring it out upon load (via merge_hashes),
just write out the one portion of the hash that we intend to save into
each unique file.  That way, there's never any collisions upon loading
all the individual files.  If there ''is'' a collision, then someone's
running MTT unsafely, so print an error message indicating that and
then abort.

This commit was SVN r1362.