When building with Visual Studio, the output is...
When building with Visual Studio, the output is huge, and errors are embedded in the output. Now we use a regex to filter out the useful messages, i.e. compilation errors.

Update the do_step command line for windows, in order to be compatible with older version of VS.

This commit was SVN r1432.