Fix tests for have_hostlist() and have_ini_maxp...
Fix tests for have_hostlist() and have_ini_maxprocs(): ensure that not
only were the values set, but that they were set to a non-empty
value.  Without these checks, then an uncommented line "max_np=" in
the INI file would wrongly set an NP value of "" (i.e., the empty string).

This commit was SVN r1616.