DoCommand: replace obscure perl code with easier-to-grok version
The descendant_processes() and flatten() codes were written in a
nicely efficient yet very difficult to understand manner.  This commit
replaces these two subs with admittedly less efficient but much easier
to understand code.

The new code also saves the first token of the command for each PID so
that it can be used in Debug/Verbose statements (e.g., identify the
command that is being killed, etc.).