Add Test Run "corefile_size_limit" param
If set to >= 0, setrlimit to limit corefiles to that size.  If set to
a negative number, corefile sizes are unlimited.  The default is
unlimited corefiles.

This functionality requires the BSD::Resource perl module.  If you
don't have it (and we therefore can't set corefile size limits),
you'll get an error (because it would be worse to generate gigabytes /
terrabytes of corefiles when you specifically asked us not to).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>