Fixes Issue #513: Compiler version truncated to fit in database variable
The compiler version was sometimes longer than 64 bytes (depending on the version, sometimes copyright information and other types of information were also included).
Truncating the compiler version to be max 64 bytes allows it to always fit within the database while still capturing enough information to determine compiler version.

Another thing that was fixed was that exit_value was sometimes not logged to DB. I fixed this by adding some more conditionals to ensure that exit_value is always logged.

Signed-off-by: Richard Barella <richard.t.barella@intel.com>