Add tables to DB for BIOS, FW, Provisioning, and Harassers
Good example of the effort it takes to make database changes while keeping compatibility with the python client.

Tables are added for the BIOS, FW, and Provisioning stages, along with Harasser table, and are connected to the Test Run table. There is no Harasser stage, and Harasser framework is not developed yet, however I included it in the DB anyways.

Signed-off-by: Richard Barella <richard.t.barella@intel.com>