Shell plugin now doesn't need a parent and can execute anywhere
If no parent is specified, Shell plugin will now use
the scratch directory as the default location to
execute content in.

This allows Shell plugin to be executed in any stage

Shell plugin checks if "asis_target" exists to determine whether to execute.
If "asis_target" is not specified, Shell plugin checks to see if it's execution
location exists to determine whether to execute. If both "asis_target"
and execution location do not exist, then the Shell plugin is executed.

Shell plugin now creates a directory for itself if it is
used in the TestGet stage

Shell plugin only executes MPIVersion checker if it is in
MiddlewareBuild or TestBuild stage

Signed-off-by: rtbarell <richard.t.barella@intel.com>