Implemented better check for Autotools.py builds.
This affects both the MiddlewareBuild and TestBuilds using Autotools.py. Instead of checking for the presence of a config file a new file was added to the builds called "build_complete" at the very end of the build process. If ASIS is requested Autotools now checks for the presence of the "build_complete" file.
fixes #745

Signed-off-by: Deb Rezanka <rezanka@lanl.gov>