Fix for using ompi with --no-shell and srun
Added support for ompi with --no-shell using srun in the SLURM.py
plugin. Tests were running on front end instead of compute nodes when
using salloc with --no-shell command. Used when running on large HPC
systems such as at NERSC where it is more efficient to get a single
SLURM allocation and run the tests within that allocation

Signed-off-by: Deb Rezanka <rezanka@lanl.gov>