Update the IUDatabase reporter
Sadly, there is no way to pass the test run summary of number passed,
failed, skipped, and timed out. However, we ensure that the coding is
now correct so the server can recompute those numbers, and hopefully we
can someday update the server to take our counts and avoid having the
concern about mismatches in categories.

Signed-off-by: Ralph Castain <rhc@pmix.org>