Check for the existence of pip
Allow user to specify multiple variations of the "pip" command (e.g.,
"pip3 pip") in case they have something set in the environment. Use the
first one that is found - so the user needs to ensure that "pip3" comes
first on their list if they are using Python 3.x and the default "pip"
is for Python 2.x.

Signed-off-by: Ralph Castain <rhc@pmix.org>